Move web and apps suites to their own folders
