+++ .github/workflows/e2e-tests-run-on-pr.yml
@@ -9,32 +9,32 @@ jobs:
     name: Run e2e tests
     runs-on: ubuntu-latest
     steps:
-    - uses: actions/checkout@v2
-    - uses: actions/setup-node@v2
-    - name: Install Dependencies
-      run: yarn install --frozen-lockfile
-    - name: Read only tests
-      uses: mujo-code/puppeteer-headful@master
-      id: test-code
-      env:
-        CI: 'true'
-        TESTING_ENV: PR
-        PR: ${{ github.event.client_payload.pr_number }}
-      with:
-        args: yarn run test-ci read_only
-    - name: Format output
-      if: ${{ always() }}
-      run: yarn run format-tests-output
-    - name: Set env variable
-      if: ${{ always() }}
-      id: format
-      run: echo "TEST_RESULTS="$(cat output.txt) >> $GITHUB_ENV
-    ## Dispatch job result to safe-react
-    - name: Repository Dispatch Comment
-      if: ${{ always() }}
-      uses: peter-evans/repository-dispatch@v1
-      with:
-        token: ${{ secrets.E2E_RUN_HOOK_ACCESS_TOKEN }}
-        repository: gnosis/safe-react
-        event-type: comment-pr-on-success
-        client-payload: '{"pr_number": "${{ github.event.client_payload.pr_number }}", "status": "${{ steps.test-code.outcome }}", "run_number": "${{ github.run_id }}", "message": "${{ env.TEST_RESULTS }}" }'
+      - uses: actions/checkout@v2
+      - uses: actions/setup-node@v2
+      - name: Install Dependencies
+        run: yarn install --frozen-lockfile
+      - name: Read only tests
+        uses: mujo-code/puppeteer-headful@master
+        id: test-code
+        env:
+          CI: 'true'
+          TESTING_ENV: PR
+          PR: ${{ github.event.client_payload.pr_number }}
+        with:
+          args: yarn run test-ci safe-web/read_only
+      - name: Format output
+        if: ${{ always() }}
+        run: yarn run format-tests-output
+      - name: Set env variable
+        if: ${{ always() }}
+        id: format
+        run: echo "TEST_RESULTS="$(cat output.txt) >> $GITHUB_ENV
+      ## Dispatch job result to safe-react
+      - name: Repository Dispatch Comment
+        if: ${{ always() }}
+        uses: peter-evans/repository-dispatch@v1
+        with:
+          token: ${{ secrets.E2E_RUN_HOOK_ACCESS_TOKEN }}
+          repository: gnosis/safe-react
+          event-type: comment-pr-on-success
+          client-payload: '{"pr_number": "${{ github.event.client_payload.pr_number }}", "status": "${{ steps.test-code.outcome }}", "run_number": "${{ github.run_id }}", "message": "${{ env.TEST_RESULTS }}" }'
