+++ src/safe-web/read_only/load_safe.test.js
@@ -9,13 +9,13 @@ import {
   clickElement,
   getInnerText,
   isTextPresent,
-} from '../../utils/selectorsHelpers'
-import { accountsSelectors } from '../../utils/selectors/accounts'
-import { generalInterface } from '../../utils/selectors/generalInterface'
-import { loadSafeForm } from '../../utils/selectors/loadSafeForm'
-import { initWithWalletConnected } from '../../utils/testSetup'
-import config from '../../utils/config'
-import { errorMsg } from '../../utils/selectors/errorMsg'
+} from '../../../utils/selectorsHelpers'
+import { accountsSelectors } from '../../../utils/selectors/accounts'
+import { generalInterface } from '../../../utils/selectors/generalInterface'
+import { loadSafeForm } from '../../../utils/selectors/loadSafeForm'
+import { initWithWalletConnected } from '../../../utils/testSetup'
+import config from '../../../utils/config'
+import { errorMsg } from '../../../utils/selectors/errorMsg'
 
 /*
 Load safe
