+++ src/safe-web/send_funds.test.js
@@ -1,5 +1,5 @@
-import { approveAndExecuteWithOwner } from '../utils/actions/approveAndExecuteWithOwner'
-import { verifySuccessfulExecution } from '../utils/actions/verifySuccesfulExecution'
+import { approveAndExecuteWithOwner } from '../../utils/actions/approveAndExecuteWithOwner'
+import { verifySuccessfulExecution } from '../../utils/actions/verifySuccesfulExecution'
 import {
   assertElementPresent,
   assertTextPresent,
@@ -10,15 +10,15 @@ import {
   getNumberInString,
   getInnerText,
   openDropdown,
-} from '../utils/selectorsHelpers'
-import { assetsTab } from '../utils/selectors/assetsTab'
-import { generalInterface } from '../utils/selectors/generalInterface'
-import { sendFundsForm } from '../utils/selectors/sendFundsForm'
-import { transactionsTab } from '../utils/selectors/transactionsTab'
-import { errorMsg } from '../utils/selectors/errorMsg'
-import { initWithDefaultSafeDirectNavigation } from '../utils/testSetup'
-import config from '../utils/config'
-import { rejectPendingTxs } from '../utils/actions/rejectPendingTxs'
+} from '../../utils/selectorsHelpers'
+import { assetsTab } from '../../utils/selectors/assetsTab'
+import { generalInterface } from '../../utils/selectors/generalInterface'
+import { sendFundsForm } from '../../utils/selectors/sendFundsForm'
+import { transactionsTab } from '../../utils/selectors/transactionsTab'
+import { errorMsg } from '../../utils/selectors/errorMsg'
+import { initWithDefaultSafeDirectNavigation } from '../../utils/testSetup'
+import config from '../../utils/config'
+import { rejectPendingTxs } from '../../utils/actions/rejectPendingTxs'
 
 /*
 Send funds
